support various channel widths <= 64

+++ src/main/cc/simif_catapult.cc
@@ -9,10 +9,10 @@ simif_catapult_t::~simif_catapult_t() {
   catapult_finish();
 }
 
-void simif_catapult_t::write(size_t addr, uint32_t data) {
+void simif_catapult_t::write(size_t addr, uint64_t data) {
   catapult_softreg_write(addr, data);
 }
 
-uint32_t simif_catapult_t::read(size_t addr) {
+uint64_t simif_catapult_t::read(size_t addr) {
   return catapult_softreg_read(addr);
 }
